Implemented commands struct
