+++ rust/src/constants.rs
@@ -1,0 +1,4 @@
+pub const DEFAULT_WIDTH: u8 = 12;
+pub const BOARD_SIZE: usize = 61;
+pub const SIZES: &str = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890-.";
+pub const DIAGONAL: u8 = 9;
