+++ rust/src/hanb.rs
@@ -1,0 +1,176 @@
+use std::cmp::min;
+
+use crate::constants::{SIZES, BOARD_SIZE};
+
+/// Hanb cell. A cell has a fixed value and can resolve to a dynamic board.
+pub struct Cell {
+    /// The value of the cell.
+    pub value: char,
+    /// The board that this cell resolves to.
+    pub board: Option<Board>,
+}
+
+impl Cell {
+    /// Create a new cell with a value.
+    pub fn new(value: char) -> Self {
+        Self { value, board: None }
+    }
+
+    /// Define the board that this cell resolves to.
+    pub fn set_board(&mut self, board: Board) {
+        self.board = Some(board);
+    }
+
+    /// Define the board that this cell resolves to from a string.
+    pub fn set_board_from_str(&mut self, board: &str) -> Result<(), String> {
+        let board = Board::new(board)?;
+        self.set_board(board);
+        Ok(())
+    }
+
+    /// Get the board that this cell resolves to.
+    pub fn get_board(&self) -> Option<&Board> {
+        self.board.as_ref()
+    }
+}
+
+/// Hanb board. A board is a collection of 61 cells + 3 optional characters that are the board
+/// color.
+pub struct Board {
+    /// The cells of the board.
+    pub cells: Vec<Cell>,
+    /// The optional characters of the board.
+    pub color: String,
+}
+
+impl Board {
+    /// Preprocess input characters in a board
+    fn preprocess(input: &str) -> String {
+        let mut output = String::new();
+        for c in input.chars() {
+            if SIZES.contains(c) {
+                output.push(c);
+            } else {
+                output.push(match c {
+                    ' ' => '-',
+                    // TODO what to do with other invalid characters?
+                    _ => '.',
+                });
+            }
+        }
+        output
+    }
+
+    /// Create a new board from a BOARD_SIZE long string. Can optionally include a color.
+    pub fn new(board: &str) -> Result<Self, String> {
+        if board.len() < 1 {
+            return Err(format!(
+                "Board must be at least {} characters long. You entered {}",
+                1,
+                board.len()
+            ));
+        }
+        let mut truncated_board = Board::preprocess(&board[0..min(board.len(), BOARD_SIZE + 3)]);
+        // Complete the remaining with '.'
+        truncated_board += &".".repeat(BOARD_SIZE + 3 - board.len());
+        let mut cells = Vec::new();
+        for c in truncated_board.chars().take(BOARD_SIZE) {
+            cells.push(Cell::new(c));
+        }
+        let color = truncated_board.chars().skip(BOARD_SIZE).take(3).collect();
+        Ok(Board { cells, color })
+    }
+
+    /// Converts the board to a string.
+    pub fn to_string(&self) -> String {
+        let mut board = String::new();
+        for cell in &self.cells {
+            board.push(cell.value);
+        }
+        board.push_str(&self.color);
+        board
+    }
+}
+
+/// Implement the board navigator that keeps track of the current level and board.
+pub struct Navigator {
+    pub root_board: Board,
+    pub cell_stack: Vec<usize>,
+}
+
+impl Navigator {
+    /// Create a new navigator with a board.
+    pub fn new(board: Board) -> Self {
+        Self {
+            root_board: board,
+            cell_stack: Vec::new(),
+        }
+    }
+
+    /// Get the current board.
+    pub fn current_board(&self) -> &Board {
+        let mut board = &self.root_board;
+        for cell in &self.cell_stack {
+            board = board.cells[*cell].get_board().unwrap();
+        }
+        board
+    }
+
+    /// Return to previous/upper board returning it
+    pub fn ascend(&mut self) -> Option<&Board> {
+        self.cell_stack.pop()?;
+        Some(self.current_board())
+    }
+
+    /// Go to a cell in the current board returning it
+    pub fn descend(&mut self, cell: usize) -> Result<&Board, String> {
+        self.resolve(cell)?;
+        self.cell_stack.push(cell);
+        Ok(self.current_board())
+    }
+
+    /// Resolve the cell at position `pos` in the current board.
+    pub fn resolve(&mut self, pos: usize) -> Result<&Board, String> {
+        if pos >= BOARD_SIZE {
+            return Err(format!(
+                "Position must be less than {}. You entered {}",
+                BOARD_SIZE, pos
+            ));
+        }
+        let cells = &self.current_board().cells;
+        if cells[pos].board.is_none() {
+            return Err(format!(
+                "Cell at position {} does not resolve to a board",
+                pos
+            ));
+        }
+        let cell = &cells[pos];
+        if let Some(board) = cell.get_board() {
+            Ok(board)
+        } else {
+            Err(format!(
+                "Cell at position {} does not resolve to a board",
+                pos
+            ))
+        }
+    }
+
+    /// Define the board for a cell in the current board.
+    pub fn define(&mut self, pos: usize, value: &str) -> Result<(), String> {
+        if pos >= self.current_board().cells.len() {
+            return Err(format!(
+                "Position must be less than {}. You entered {}",
+                self.current_board().cells.len(),
+                pos
+            ));
+        }
+
+        let mut board = &mut self.root_board;
+        for cell in &self.cell_stack {
+            board = board.cells[*cell].board.as_mut().unwrap();
+        }
+        board.cells[pos].set_board_from_str(value)?;
+        Ok(())
+    }
+}
+
